deletes authentication from base model. adds separate integration suite for session auth.
